\name{BroadstoneStream}
\docType{data}
\alias{BroadstoneStream}
\title{Broadstone Stream}
\description{The community of Broadstone Stream.}
\usage{BroadstoneStream}
\format{\code{Community}.}
\references{
Woodward, G. and Speirs, D.C. and Hildrew, A.G. (2005) Quantification and 
resolution of a complex, size-structured food web. Advances in Ecological 
Research \bold{36}, 85--135.
}
\keyword{datasets}

