\name{ChesapeakeBay}
\docType{data}
\alias{ChesapeakeBay}
\title{ChesapeakeBay}
\description{The community of Chesapeake Bay sampled in the years 1983 to 1986.}
\usage{ChesapeakeBay}
\format{\code{Community}.}
\source{Baird and Ulanowicz, 1989; Bersier et al, 2002.}
\references{
Bersier, L. and Banasek-Richter, C. and Cattin, M. (2002) Ecology 
\bold{80} 2394--2407.

Baird, D. and Ulanowicz. R. E. (1989) Ecological Monographs \bold{59}, 329--364.
}

\keyword{datasets}
