\name{CommunityPropertyNames}
\alias{CommunityPropertyNames}
\title{Names of community properties}
\description{Returns a vector of names of community properties.}
\usage{CommunityPropertyNames(community)}
\arguments{
  \item{community}{an object of class \code{Community}.}
}
\value{A vector of names of community properties.}

\author{Lawrence Hudson}

\seealso{\code{\link{Community}}, \code{\link{CP}}, \code{\link{CPS}}}

\examples{
data(TL84)
CommunityPropertyNames(TL84)
}
\keyword{utilities}

