\name{Millstream}
\docType{data}
\alias{Millstream}
\title{Millstream}
\description{The control and drought treatments from one of the four replicates 
from a long-running study of the effects of drought on community structure.}
\usage{Millstream}
\format{\code{CommunityCollection}.}
\references{
Ledger, M.E. and Harris, R.M.L. and Armitage, P.D. and Milner, A.M. (2008) 
Disturbance frequency influences patch dynamics in stream benthic algal 
communities. Oecologia. \bold{155}, 4, 809--819.

Ledger, M.E. and Edwards, F.K. and Brown, L.E. and Milner, A.M. and Woodward, 
G. (2011) Impact of simulated drought on ecosystem biomass production: an 
experimental test in stream mesocosms. \bold{17}, 7, 2288--2297.

Woodward, G. and Brown, L.E and Edwards, F. and Hudson, L.N. and Milner, A.M. 
and Reuman, D.C. and Mark E.L. (2012) Climate change impacts in multispecies 
systems: drought alters food web size-structure in a field experiment. 
Philosophical Transactions of the Royal Society B: Biological Sciences.
}

\keyword{datasets}

