\name{TL84}
\docType{data}
\alias{TL84}
\alias{TL86}
\title{Tuesday Lake datasets}
\description{The communities of Tuesday Lake, Michigan, USA sampled in 
1984 and 1986.}
\usage{TL84
TL86}

\format{\code{Community} objects.}
\references{
Carpenter, S.R. and Kitchell, J.F. (1996) \emph{The trophic cascade in lakes}. 
Cambridge University Press.

Cohen, J.E. and Jonsson, T. and Carpenter, S.R. (2003) Ecological community 
description using the food web, species abundance, and body size. Proceedings 
of the National Academy of Sciences of the United States of America 
\bold{100}, 4, 1781--1786.

Jonsson, T. and Cohen, J.E. and Carpenter, S.R. (2005) Food webs, body size, 
and species abundance in ecological community description. Advances in 
Ecological Research \bold{36}, 1--84.
}
\keyword{datasets}

