\name{YthanEstuary}
\docType{data}
\alias{YthanEstuary}
\title{Ythan Estuary}
\description{The community of Ythan Estuary.}
\usage{YthanEstuary}
\format{\code{Community}.}
\references{
Hall, SJ and Raffaelli, D. (1991) Food-web patterns: lessons from a 
species-rich web. Journal of Animal Ecology. \bold{60}, 3, 823--841.

Emmerson, M.C. and Raffaelli, D. (2004) Predator-prey body size, interaction 
strength and the stability of a real food web. Journal of Animal Ecology 
\bold{73}, 3, 399--409.
}
\keyword{datasets}

