\name{Benguela}
\docType{data}
\alias{Benguela}
\title{Benguela}
\description{The food-web of the Benguela ecosystem.}
\usage{Benguela}
\format{\code{Community}.}
\references{
Yodzis, P. (1988) Local trophodynamics and the interaction of marine mammals 
and fisheries in the Benguela ecosystem. Journal of Animal Ecology \bold{67}, 
4, 635--658.
}
\keyword{datasets}

