\name{BroadstoneStream}
\docType{data}
\alias{BroadstoneStream}
\title{Broadstone Stream}
\description{The community of Broadstone Stream.

Taxonomic classification provided by Guy Woodward.}
\format{\code{Community}.}
\source{Woodward et al, 2005.}
\references{
Woodward, G. and Speirs, D.C. and Hildrew, A.G. (2005) Quantification and 
resolution of a complex, size-structured food web. Advances in Ecological 
Research \bold{36}, 85--135.
}
\keyword{datasets}

