\name{pHWebs}
\docType{data}
\alias{pHWebs}
\title{pHWebs}
\description{Ten stream food webs sampled across a large pH gradient.}
\format{\code{CommunityCollection}.}
\source{Layer et al 2010.}
\references{
Layer, K. and Riede, J.O. and Hildrew, A.G. and Woodward, G. (2010) Food web 
structure and stability in 20 streams across a wide pH gradient. 
Advances in Ecological Research \bold{42}, 265--299.
}

\keyword{datasets}

