\name{SkipwithPond}
\docType{data}
\alias{SkipwithPond}
\title{SkipwithPond}
\description{The food-web of Skipwith Pond. 

Taxonomic classification provided by Guy Woodward.}
\usage{SkipwithPond}
\format{\code{Community}.}
\source{Warren, 1989.}
\references{
Warren, P.H. (1989) Spatial and temporal variation in the structureof a 
freshwater food web. Oikos \bold{55}, 299--311.
}
\keyword{datasets}

