\name{TL84}
\docType{data}
\alias{TL84}
\alias{TL86}
\title{Tuesday Lake datasets}
\description{The communities of Tuesday Lake, Michigan, USA sampled in 1984 and 
1986.

Taxonomic classification 
fish: Froese and Pauly (2012), 
invertebrates: Smith (2001), 
phytoplankton: Guiry and Guiry (2012).
}

\format{\code{Community} objects.}
\source{Carpenter and Kitchell, 1996; Cohen et al, 2003; Johnsson et al 2005.}
\references{
Carpenter, S.R. and Kitchell, J.F., eds. (1996) \emph{The trophic cascade in 
lakes}. Cambridge University Press.

Cohen, J.E. and Jonsson, T. and Carpenter, S.R. (2003) Ecological community 
description using the food web, species abundance, and body size. Proceedings 
of the National Academy of Sciences of the United States of America 
\bold{100}, 4, 1781--1786.

Froese, R. and Pauly, D., eds. 2012. FishBase. World Wide Web electronic 
publication. \url{https://www.fishbase.org}, version (08/2012).

Guiry, M.D. and Guiry, G.M. (2012) AlgaeBase. World-wide electronic 
publication, National University of Ireland, Galway. 
\url{https://www.algaebase.org}; searched on 20 September 2012.

Jonsson, T. and Cohen, J.E. and Carpenter, S.R. (2005) Food webs, body size, 
and species abundance in ecological community description. Advances in 
Ecological Research \bold{36}, 1--84.

Smith, D.G. (2001) \emph{Pennak's Freshwater Invertebrates of the United 
States: Porifera to Crustacea}, John Wiley and Sons, 4th Edition.
}
\keyword{datasets}

