\name{YthanEstuary}
\docType{data}
\alias{YthanEstuary}
\title{Ythan Estuary}
\description{The community of Ythan Estuary. 

Taxonomic classification for 
birds: Dickinson et al (2003), 
fish: Froese and Pauly (2012), 
invertebrates: Appeltans et al (2012), 
mammals: Wilson and Reeder (2005), 
phytoplankton: Guiry and Guiry (2012).
}
\usage{YthanEstuary}
\format{\code{Community}.}
\source{Hall and Raffaelli, 1991; Emmerson and Raffaelli, 2004.}
\references{
Appeltans W., Bouchet P., Boxshall G.A., De Broyer C., de Voogd N.J., 
Gordon D.P., Hoeksema B.W., Horton T., Kennedy M., Mees J., Poore G.C.B., 
Read G., St\"{o}hr S., Walter T.C., Costello M.J. Editors. (2012) World 
Register of Marine Species. Accessed at \url{http://www.marinespecies.org} on 
2012-09-20.

Dickinson, E.C., ed. (2003) The Howard and Moore complete checklist of the 
birds of the world, Christopher Helm, London, third edition.

Emmerson, M.C. and Raffaelli, D. (2004) Predator-prey body size, interaction 
strength and the stability of a real food web. Journal of Animal Ecology 
\bold{73}, 3, 399--409.

Froese, R. and Pauly, D., eds. 2012. FishBase. World Wide Web electronic 
publication. \url{www.fishbase.org}, version (08/2012).

Guiry, M.D. and Guiry, G.M. (2012) AlgaeBase. World-wide electronic 
publication, National University of Ireland, Galway. 
\url{http://www.algaebase.org}; searched on 20 September 2012.

Hall, S.J. and Raffaelli, D. (1991) Food-web patterns: lessons from a 
species-rich web. Journal of Animal Ecology. \bold{60}, 3, 823--841.

Wilson, D.E. and Reeder, D.M., eds. (2005) Mammal species of the world. A 
taxonomic and geographic reference. Johns Hopkins University Press, third 
edition.
}
\keyword{datasets}

